Use short form for attributes when possible
This will result in a different output in the case where an elixir expressions
returns `true` or `false` or `nil`:

```
<div attr="true"> vs <div attr="">
<div attr="false"> vs <div>
<div attr=""> vs <div>
```

This could be a meaningfull difference if you are passing data to javascript (e.g. `data-example="..."`), but for standard HTML the previous conversions are non-sensical (e.g. `disabled="false"` or `disabled=""` mean that it is disabled!)